server/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended
python-dotenv
pymysql
python-dotenv
werkzeug
python-dateutil